Move Vimscript grammar to engine
