+++ src/main/java/com/maddyhome/idea/vim/vimscript/parser/visitors/CommandVisitor.kt
@@ -101,15 +101,15 @@ import com.maddyhome.idea.vim.vimscript.model.expressions.Scope
 import com.maddyhome.idea.vim.vimscript.model.expressions.SimpleExpression
 import com.maddyhome.idea.vim.vimscript.model.expressions.operators.AssignmentOperator
 import com.maddyhome.idea.vim.vimscript.model.expressions.operators.AssignmentOperator.Companion.getByValue
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptBaseVisitor
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.CallCommandContext
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.DelfunctionCommandContext
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.EchoCommandContext
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.ExprContext
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.OtherCommandContext
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.RangeContext
-import com.maddyhome.idea.vim.vimscript.parser.generated.VimscriptParser.RangeOffsetContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptBaseVisitor
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.CallCommandContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.DelfunctionCommandContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.EchoCommandContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.ExprContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.OtherCommandContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.RangeContext
+import com.maddyhome.idea.vim.parser.generated.VimscriptParser.RangeOffsetContext
 import org.antlr.v4.runtime.ParserRuleContext
 import java.util.stream.Collectors
 import kotlin.reflect.KClass
