+++ vim-engine/build.gradle.kts
@@ -65,8 +65,8 @@ tasks {
 
     generateGrammarSource {
         maxHeapSize = "128m"
-        arguments.addAll(listOf("-package", "com.maddyhome.idea.vim.regexp.parser.generated", "-visitor"))
-        outputDirectory = file("src/main/java/com/maddyhome/idea/vim/regexp/parser/generated")
+        arguments.addAll(listOf("-package", "com.maddyhome.idea.vim.parser.generated", "-visitor"))
+        outputDirectory = file("src/main/java/com/maddyhome/idea/vim/parser/generated")
     }
 
     named("compileKotlin") {
